% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis_monomials.R
\name{basis_monomials}
\alias{basis_monomials}
\title{Enumerate basis monomials}
\usage{
basis_monomials(indeterminates, d)
}
\arguments{
\item{indeterminates}{a character vector}

\item{d}{the highest total degree}
}
\value{
a \code{\link[=mpolyList]{mpolyList()}} object of monomials
}
\description{
Enumerate basis monomials in the standard basis up to a given degree.
}
\examples{

basis_monomials(c("x", "y"), 2)
basis_monomials(c("x", "y"), 3)
basis_monomials(c("x", "y", "z"), 2)
basis_monomials(c("x", "y", "z"), 3)

}
