% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chebyshev.R
\name{chebyshev}
\alias{chebyshev}
\title{Chebyshev polynomials}
\usage{
chebyshev(degree, kind = 1, indeterminate = "x")
}
\arguments{
\item{degree}{degree of polynomial}

\item{kind}{1 or 2 (Chebyshev polynomials of the first and second kinds)}

\item{indeterminate}{indeterminate}
}
\value{
a mpoly object
}
\description{
Chebyshev polynomials
}
\examples{
chebyshev(0)
chebyshev(1)
chebyshev(2)
chebyshev(3)
chebyshev(4)
chebyshev(5)
chebyshev(6)
chebyshev(10)

chebyshev(0:5)
chebyshev(0:5, 2)

chebyshev(0:5, indeterminate = "t")


\dontrun{  # visualize the chebyshev polynomials

library(ggplot2); theme_set(theme_bw())
library(reshape2)

s <- seq(-1, 1, length.out = 201)
N <- 5 # number of chebyshev polynomials to plot
(chebPolys <- chebyshev(0:N))

df <- t(sapply(s, as.function(chebPolys)) )
df <- as.data.frame(cbind(s, df))
names(df) <- c("x", paste0("T_", 0:N))
mdf <- melt(df, id = "x")
qplot(x, value, data = mdf, geom = "line", color = variable)


}
}
\author{
David Kahle modifying code from Adam Leerich modifying code from Hadley Wickham
}

