% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_to_mp.R
\name{eq_to_mp}
\alias{eq_to_mp}
\title{Convert an equation to a polynomial}
\usage{
eq_to_mp(string, ...)
}
\arguments{
\item{string}{a character string containing a polynomial, see examples}

\item{...}{arguments to pass to \code{\link[mpoly::mpoly]{mpoly()}}}
}
\value{
An object of class mpoly or mpolyList.
}
\description{
Convert characters of the form \code{"p1 = p2"} (or similar) to an mpoly
object representing \code{p1 - p2}.
}
\examples{

eq_to_mp(c("y = x", "y ==  (x + 2)"))



}
\seealso{
\code{\link{mpoly}}
}
\author{
David Kahle \email{david@kahle.io}
}
