% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponents.R
\name{exponents}
\alias{exponents}
\title{Extract exponents}
\usage{
exponents(x, reduced = FALSE)
}
\arguments{
\item{x}{an \code{\link{mpoly}} object}

\item{reduced}{if TRUE, don't include zero degrees}
}
\value{
a list of named vectors of degrees
}
\description{
Extract exponents of a multivariate polynomial.
}
\examples{
x <- mp("x^4 + y + 2 x y^2 - 3 z")
exponents(x)
exponents(x, reduce = TRUE)
lapply(exponents(x), is.integer)

}

