% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract}
\alias{LC}
\alias{LM}
\alias{LT}
\alias{[.mpoly}
\alias{extract}
\alias{multideg}
\title{Polynomial extraction}
\usage{
\method{[}{mpoly}(x, ndx)

LT(x, varorder = vars(x), order = "lex")

LC(x, varorder = vars(x), order = "lex")

LM(x, varorder = vars(x), order = "lex")

multideg(x, varorder = vars(x), order = "lex")
}
\arguments{
\item{x}{an object of class mpoly}

\item{ndx}{a subsetting index}

\item{varorder}{the order of the variables}

\item{order}{a total order used to order the terms}

\item{...}{additional arguments}
}
\value{
An object of class mpoly
}
\description{
Extract components of a multivariate polynomial
}
\examples{
(p <- mp("x y + x (x+1) (x+2) x z + 3 x^10"))
p[2]
LT(p)
LC(p)
LM(p)
multideg(p)
}

