% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvariety.R
\name{ggvariety}
\alias{ggvariety}
\title{Plot a variety}
\usage{
\method{ggvariety}{mpoly}(mp, xlim, ylim, obs, ...)
}
\arguments{
\item{mp}{an mpoly object}

\item{xlim}{vector representing x bounds}

\item{ylim}{vector representing y bounds}

\item{obs}{number of ??????}

\item{...}{additional parameters}
}
\value{
A ggplot object containing variety plot
}
\description{
Use geom_contour() and ggplot() to plot an mpoly 
object representing a variety.
}
\examples{

# Simple linear variety
variety <- mp("y-x+2")
ggvariety(variety, c(-2,2), c(-2,2), 20)

# Alpha curve (not smooth)
ggvariety(mp("y^2-x^3-x^2"), c(-2,2), c(-2,2), 50)

# Alpha curve (smooth)
ggvariety(mp("y^2-x^3-x^2"), c(-2,2), c(-2,2), 400)


}
