% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grobner.R
\name{grobner}
\alias{grobner}
\title{REMOVED -- Compute a grobner basis of a list of multivariate 
polynomials.}
\usage{
grobner(mpolyList, varorder = vars(mpolyList), order = "lex")
}
\arguments{
\item{mpolyList}{an object of class mpolyList}

\item{varorder}{order of variables}

\item{order}{total order to be considered for monomials (e.g. 
lexicographic)}
}
\value{
An object of class mpolyList.
}
\description{
This function has been removed to eliminate mpoly's dependence on
packages that only it uses.  To compute a Grobner basis of a 
collection of multivariate polynomials, checkout the new m2r
package, which you can download with the code in the first example.
}
\details{
grobner computes a Grobner basis for a collection of multivarite 
polynomials represented as an object of class mpolyList. Note 
that the polynomials printed after calculation are unlikely to be
properly ordered; this is because the order of the monomials 
displayed is governed by the print method, not the mpoly's 
themselves.
}
\examples{
\dontrun{


# code to download m2r:
# note that to do this you should have Macaulay2 installed,
# see https://github.com/musicman3320/m2r and 
# http://www.math.uiuc.edu/Macaulay2/Downloads/
if(!require(devtools)) install.packages("devtools")
devtools::install_github("musicman3320/m2r")





}
}
