% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homogenize.R
\name{homogenize}
\alias{dehomogenize}
\alias{homogeneous_components}
\alias{homogenize}
\alias{is.homogeneous}
\title{Homogenize a polynomial}
\usage{
homogenize(x, var = "t")

dehomogenize(x, var)

is.homogeneous(x)

homogeneous_components(x)
}
\arguments{
\item{x}{an \code{\link{mpoly}} object}

\item{var}{name of homogenization}
}
\value{
a (homogenized) mpoly
}
\description{
Homogenize a polynomial.
}
\examples{

x <- mp("x^4 + y + 2 x y^2 - 3 z")
is.homogeneous(x)
(xh <- homogenize(x))
is.homogeneous(xh)

homogeneous_components(x)

homogenize(x, "o")

xh <- homogenize(x)
dehomogenize(xh, "t")

\dontrun{ demonstrates error
xh <- mp("x^2 t + y")
is.homogeneous(xh)
dehomogenize(xh, "t")
}

}

