% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/predicates.R
\name{is.constant}
\alias{is.constant}
\title{Determine if a polynomial is a constant}
\usage{
is.constant(p)
}
\arguments{
\item{x}{object to be tested}
}
\value{
Vector of logicals.
}
\description{
Test whether an mpoly is a constant polynomial or the elements of an mpolyList are constant polynomials.
}
\examples{
p <- mp("5")
is.mpoly(p)
is.constant(p)

is.constant(mp(c("x + 1", "7", "y - 2")))
}

