% Generated by roxygen2 (4.0.2): do not edit by hand
\name{is.mpolyList}
\alias{is.mpolyList}
\title{Test whether an object is an mpolyList.}
\usage{
is.mpolyList(mpolyList)
}
\arguments{
\item{mpolyList}{object whose class is in question}
}
\value{
Logical.
}
\description{
Test whether an object is an mpolyList.
}
\examples{
is.mpolyList(mp(c('x + 1', 'y - 1')))
}

