% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/is.wholenumber.R
\name{is.wholenumber}
\alias{is.wholenumber}
\title{Test whether an object is a whole number.}
\usage{
is.wholenumber(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{object to be tested}

\item{tol}{tolerance within which a number is said to be whole}
}
\value{
Vector of logicals.
}
\description{
Test whether an object is a whole number.
}
\examples{
is.wholenumber(seq(-3,3,.5))
}

