% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCM.R
\name{LCM}
\alias{LCM}
\title{Compute the least common multiple of two numbers.}
\usage{
LCM(x, y)
}
\arguments{
\item{x}{an object of class numeric}

\item{y}{an object of class numeric}
}
\value{
The least common multiple of x and y.
}
\description{
A simple algorithm to compute the least common multiple of two 
numbers
}
\examples{
LCM(5,7)
LCM(5,8)
LCM(5,9)
LCM(5,10)
Reduce(LCM, 1:10) # -> 2520
}

