% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mp}
\alias{mp}
\title{Define a multivariate polynomial.}
\usage{
mp(string, varorder)
}
\arguments{
\item{string}{a character string containing a polynomial, see examples}

\item{varorder}{(optional) order of variables in string}
}
\value{
An object of class mpoly.
}
\description{
mp is a smart function which attempts to create a formal mpoly object from a character string containing the usual representation  of a multivariate polynomial.
}
\examples{
( m <- mp('x + y + x y') )
is.mpoly( m )
unclass(m)





gradient( mp("x + 2 y + x^2 y + x y z") )
gradient( mp("(x + y)^10") )

# mp and the print methods are kinds of inverses of each other
( polys <- mp(c('x + y', 'x - y')) )
strings <- print(polys)
strings
mp(strings)




















# working new in version 0.6
mp('x + 4x') # note the 4x as opposed to 4 x
mp('x - 4x') # same
mp('x -1') # -> -1 x
mp("x(1+x)")

mp("x ((x + y) + 2)")
mp("1+x")
}
\author{
David Kahle \email{david.kahle@gmail.com}
}
\seealso{
\code{\link{mpoly}}
}

