% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpoly-defunct.R
\name{mpoly-defunct}
\alias{mpoly-defunct}
\title{Defunct mpoly functions}
\description{
This is a list of past functions of the mpoly package.
}
\details{
The following are defunct mpoly functions:

\itemize{

\item grobner

}
}
