% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpolyArithmetic.R, R/mpolyListArithmetic.R
\name{mpolyArithmetic}
\alias{mpolyArithmetic}
\alias{+.mpoly}
\alias{-.mpoly}
\alias{*.mpoly}
\alias{^.mpoly}
\alias{^.mpolyList}
\title{Arithmetic with multivariate polynomials}
\usage{
\method{+}{mpoly}(e1, e2)

\method{-}{mpoly}(e1, e2)

\method{*}{mpoly}(e1, e2)

\method{^}{mpoly}(e1, e2)

\method{^}{mpolyList}(e1, e2)
}
\arguments{
\item{e1}{an object of class mpoly}

\item{e2}{an object of class mpoly}
}
\value{
object of class mpoly
}
\description{
Arithmetic with multivariate polynomials
}
\examples{

p <- mp("x + y")
p + p
p - p
p * p
p^2
p^10


mp("(x+1)^10")
p + 1
2*p



}
