% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mpolyExp.R
\name{mpolyExp}
\alias{^.mpoly}
\alias{mpolyExp}
\title{Exponentiate multivariate polynomials.}
\usage{
\method{^}{mpoly}(e1, e2)
}
\arguments{
\item{e1}{an object of class mpoly}

\item{e2}{a positive integer}
}
\value{
object of class mpoly
}
\description{
Exponentiate multivariate polynomials.
}
\examples{
list <- list(
  c(x = 1, coef = 1),
  c(y = 1, coef = -4),
  c(x = 2, z = 1, coef = -3),
  c(t = 3, y = 4, coef = 2)
)
mp1 <- mpoly(list)
mp1 * mp1
mp1^2
mp1^10
}

