% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mpolyListAdd.R
\name{mpolyListAdd}
\alias{+.mpolyList}
\alias{mpolyListAdd}
\title{Add vectors multivariate polynomials element-wise.}
\usage{
\method{+}{mpolyList}(e1, e2)
}
\arguments{
\item{e1}{an object of class mpolyList}

\item{e2}{an object of class mpolyList}
}
\value{
An object of class mpolyList.
}
\description{
Compute the sum of two mpolyList objects.
}
\examples{
( ms1 <- mp( c('x + 1', 'x + 2') ) )
( ms2 <- mp( c('x + 1', 'y + 2') ) )
ms1 + ms2
}

