% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpolyListArithmetic.R
\name{mpolyListArithmetic}
\alias{mpolyListArithmetic}
\alias{+.mpolyList}
\alias{-.mpolyList}
\alias{*.mpolyList}
\title{Element-wise arithmetic with vectors of multivariate polynomials.}
\usage{
\method{+}{mpolyList}(e1, e2)

\method{-}{mpolyList}(e1, e2)

\method{*}{mpolyList}(e1, e2)
}
\arguments{
\item{e1}{an object of class mpolyList}

\item{e2}{an object of class mpolyList}
}
\value{
An object of class mpolyList.
}
\description{
Element-wise arithmetic with vectors of multivariate polynomials.
}
\examples{

( ms1 <- mp( c('x + 1', 'x + 2') ) )
( ms2 <- mp( c('x + 1', 'y + 2') ) )
ms1 + ms2
ms1 - ms2
ms1 * ms2

}
