% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mpolyListProd}
\alias{*.mpolyList}
\alias{mpolyListProd}
\title{Multiply vectors multivariate polynomials element-wise.}
\usage{
\method{*}{mpolyList}(e1, e2)
}
\arguments{
\item{e1}{an object of class mpoly}

\item{e2}{an object of class mpoly}
}
\value{
An object of class mpoly.
}
\description{
Compute the product of two mpolyList objects.
}
\examples{
( ms1 <- mp( c('x + 1', 'x + 2') ) )
( ms2 <- mp( c('x - 1', 'x - 2') ) )
ms1 * ms2
# mpoly * mpolyList not available.
}

