% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mpolyListAdd.R
\name{mpolyListSub}
\alias{-.mpolyList}
\alias{mpolyListSub}
\title{Subtract vectors of multivariate polynomials element-wise.}
\usage{
\method{-}{mpolyList}(e1, e2)
}
\arguments{
\item{e1}{an object of class mpolyList}

\item{e2}{an object of class mpolyList}
}
\value{
An object of class mpolyList.
}
\description{
Compute the difference of two mpolyList objects.
}
\examples{
( ms1 <- mp( c('x + 1', 'x + 2') ) )
( ms2 <- mp( c('x + 1', 'y + 2') ) )
ms1 - ms2
}

