% Generated by roxygen2 (4.0.2): do not edit by hand
\name{parse_nonparenthetical_polynomial}
\alias{parse_nonparenthetical_polynomial}
\title{string <- "-4 + 2+2 x +   1 x y^4 -3 prq^3 -y - 3 x 2 - 3 y -2"
parse_nonparenthetical_polynomial(string)
parse_nonparenthetical_polynomial("x    +       y")}
\usage{
parse_nonparenthetical_polynomial(string)
}
\description{
string <- "-4 + 2+2 x +   1 x y^4 -3 prq^3 -y - 3 x 2 - 3 y -2"
parse_nonparenthetical_polynomial(string)
parse_nonparenthetical_polynomial("x    +       y")
}

