% Generated by roxygen2 (4.0.2): do not edit by hand
\name{parse_nonparenthetical_term}
\alias{parse_nonparenthetical_term}
\title{parse_nonparenthetical_term("12var 2 y 2x")
parse_nonparenthetical_term("2      -2")
parse_nonparenthetical_term("2 x y^2 3 -1           3^2")
parse_nonparenthetical_term("x")}
\usage{
parse_nonparenthetical_term(string)
}
\description{
parse_nonparenthetical_term("12var 2 y 2x")
parse_nonparenthetical_term("2      -2")
parse_nonparenthetical_term("2 x y^2 3 -1           3^2")
parse_nonparenthetical_term("x")
}

