% Generated by roxygen2 (4.0.2): do not edit by hand
\name{parse_parenthetical_term}
\alias{parse_parenthetical_term}
\title{parse_parenthetical_term(" 3 (x + y) 4 (x - y) ")
parse_parenthetical_term("(x + y) (x - y)")
parse_parenthetical_term("(x + y)")
parse_parenthetical_term("((5^2))")}
\usage{
parse_parenthetical_term(string)
}
\description{
parse_parenthetical_term(" 3 (x + y) 4 (x - y) ")
parse_parenthetical_term("(x + y) (x - y)")
parse_parenthetical_term("(x + y)")
parse_parenthetical_term("((5^2))")
}

