% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\docType{package}
\name{Rcppmpoly}
\alias{Rcppmpoly}
\alias{package-Rcppmpoly}
\alias{Rcppmpoly-package}
\title{Multivariate polynomials in R.}
\description{
A package for symbolic computation and more with multivariate polynomials
}
