% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_unipoly.R
\name{solve_unipoly}
\alias{solve_unipoly}
\title{Solve a univariate mpoly with polyroot}
\usage{
solve_unipoly(mpoly, real_only = FALSE)
}
\arguments{
\item{mpoly}{an mpoly}

\item{real_only}{return only real solutions?}
}
\description{
Solve a univariate mpoly with polyroot
}
\examples{

solve_unipoly(mp("x^2 - 1")) # check x = -1, 1
solve_unipoly(mp("x^2 - 1"), real_only = TRUE)

}
