% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/terms.mpoly.R
\name{terms.mpoly}
\alias{terms.mpoly}
\title{Extract the terms of a multivariate polynomial.}
\usage{
\method{terms}{mpoly}(x, ...)
}
\arguments{
\item{x}{an object of class mpoly}

\item{...}{additional parameters}
}
\value{
An object of class mpolyList.
}
\description{
Compute the terms of an mpoly object as a mpolyList.
}
\examples{
terms(mp('x^2 - y + x y z'))
is.mpolyList( terms(mp('x^2 - y + x y z')) )
str(unclass( terms(mp('x^2 - y + x y z')) ))
}

