% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars.R
\name{vars}
\alias{vars}
\title{Determine the variables in a mpoly object.}
\usage{
vars(p)
}
\arguments{
\item{p}{An mpoly or mpolyList object.}
}
\value{
A character vector of the variable names.
}
\description{
Determine the variables in a mpoly object.
}
\examples{

p <- mp("x + y^2")
vars(p)

p <- mp(c("x + y^2", "y - 2 x"))
vars(p)

}
