% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bezier.R
\name{bezier}
\alias{bezier}
\title{Bezier polynomials}
\usage{
bezier(..., indeterminate = "t")
}
\arguments{
\item{...}{either a sequence of points or a matrix/data frame of points, see examples}

\item{indeterminate}{the indeterminate of the resulting polynomial}
}
\value{
a mpoly object
}
\description{
Compute the Bezier polynomials of a given collection of points.  Note that
evaluating the polynomials with \code{\link{as.function.mpoly}} will not take
advantage of efficient algorithms such as De Casteljau's algorithm to do so.
For that, see the bezier package.
}
\examples{
p1 <- c(0,  0)
p2 <- c(1,  1)
p3 <- c(2, -1)
p4 <- c(3,  0)
bezier(p1, p2, p3, p4)


points <- data.frame(x = 0:3, y = c(0,1,-1,0))
bezier(points)


points <- data.frame(x = 0:2, y = c(0,1,0))
bezier(points)



\dontrun{  # visualize the bernstein polynomials

library(ggplot2); theme_set(theme_bw())

s <- seq(0, 1, length.out = 101)

## example 1
points <- data.frame(x = 0:3, y = c(0,1,-1,0))
(bezPolys <- bezier(points))

df <- t(sapply(s, as.function(bezPolys)) )
df <- as.data.frame(df)
names(df) <- c("x", "y")
qplot(x, y, data = df, geom = "path") +
  geom_path(data = points, color = "red") +
  geom_point(data = points, color = "red", size = 8)




## example 2
points <- data.frame(x = 0:2, y = c(0,1,0))
(bezPolys <- bezier(points))

df <- t(sapply(s, as.function(bezPolys)) )
df <- as.data.frame(df)
names(df) <- c("x", "y")
qplot(x, y, data = df, geom = "path") +
  geom_path(data = points, color = "red") +
  geom_point(data = points, color = "red", size = 8)




## example 3
points <- data.frame(x = c(-1,-2,2,1), y = c(0,1,1,0))
(bezPolys <- bezier(points))

df <- t(sapply(s, as.function(bezPolys)) )
df <- as.data.frame(df)
names(df) <- c("x", "y")
qplot(x, y, data = df, geom = "path") +
  geom_path(data = points, color = "red") +
  geom_point(data = points, color = "red", size = 8)



## example 4
points <- data.frame(x = c(-1,2,-2,1), y = c(0,1,1,0))
(bezPolys <- bezier(points))

df <- t(sapply(s, as.function(bezPolys)) )
df <- as.data.frame(df)
names(df) <- c("x", "y")
qplot(x, y, data = df, geom = "path") +
  geom_path(data = points, color = "red") +
  geom_point(data = points, color = "red", size = 8)




}
}
\author{
David Kahle
}

