% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deriv.mpoly.R
\name{gradient}
\alias{gradient}
\title{Compute gradient of a multivariate polynomial.}
\usage{
gradient(mpoly)
}
\arguments{
\item{mpoly}{an object of class mpoly}
}
\value{
An object of class mpoly or mpolyList.
}
\description{
This is a wrapper for deriv.mpoly.
}
\examples{
m <- mp('x y + y z + z^2')
gradient(m)
}
\seealso{
\code{\link{deriv.mpoly}}
}

