% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/is.mpoly.R
\name{is.mpoly}
\alias{is.mpoly}
\title{Test whether an object is an mpoly object.}
\usage{
is.mpoly(x)
}
\arguments{
\item{x}{object to be tested}
}
\value{
Vector of logicals.
}
\description{
Test whether an object is an mpoly object.
}

