% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars.R
\name{vars}
\alias{vars}
\title{Determine the variables in a mpoly object.}
\usage{
vars(mpoly)
}
\arguments{
\item{mpoly}{an object of class mpoly}
}
\value{
A character vector of the variable names.
}
\description{
Determine the variables in a mpoly object.
}
\examples{
list <- list(
  c(x = 5, coef = 2), 
  c(y = 2, coef = -3), 
  c(x = 1, y = 3, coef = 1)
)
p <- mpoly(list)
vars(p)
}
